remove dead code and ensure floating-point division
